Added CXF 2.x Web Service Client test
CXF 2.x is Web Service Runtime so there is a new class WebServiceRuntime
which is used to distinguist between CXF 2.x and JBoss WS runtime.
It's downloaded via maven download plugin and path to it is stored into
properties file ws.properties.

Tests for WsClient was extracted to WsClientTestTemplate which
covers testing. It's a super class for CxfWsClientTest and
WsClientTest, which specify the WebServiceRuntime.

Adding choice of Web Service Runtime, which is used to create dynamic web
project, brings problems with different dialogs layout and also need of having
more informations (like name of ear project).

There is also huge refactoring. This was also opportunity to reimplement
methods and implement new ones using Red Deer API.