Webservices: Splitted and updated test case testing WS Tester prompting values
Previously, there was only one complex test which always failed in the middle
of execution due to bug with default values. Because of that, the test case
wasn't able to report if there would be broken something in the second half of
the test case. To provide better testing, the single complex test case was
splitted into three test cases. This change also brings opportunity to write
more test cases.

Also another change is with mandatory values. The name has a star notifying
that the value is mandatory.

Added RedDeer WaitWhile to reduce unnecessary waiting for shell with text.