Webservices: Created test cases for PathAnnotationSupportTest and fields
PathAnnotationSupportTest contained tests testing RESTful explorer
but not tests testing @PathParam. So the tests from PathAnnotationSupportTest
class were moved to RESTfulExplorerTest class and new test cases testing
@PathParam were created.

RESTfulExplorerTest contained test for RESTful Support so it was
moved to a new class RESTfulSupportTest

The class PathAnnotationSupportTest was renamed to
PathParamAnnotationSupportTest because it's testing @PathParam and not @Path
annotation.

Another classes QueryAnnotationSupportTest and MatrixAnnotationSupportTest
were also renamed to have "Param" in the name.

Created test cases testing fields annotated with @QueryParam and
also @MatrixParam.