Unify the use of swtbot.test.skip
The property swtbot.test.skip is set in the jbt parent pom.
in the past most, but not all, tests used it to configure
<skip> for tycho-surefire-plugin. Now I moved this config to
tests/pom.xml so that it's consistently used everywhere.
At the same time, I changed maven-download-plugin definitions
to download-maven-plugin where applicable - this is the new artifactID.
And I set up the skip there as well. Same for maven-dependency-plugin.