WebServices: Reflected change of Simple Web Service dialog
Removed SimpleRESTWebServiceTest because Simple Web Service Dialog can't be
used to create JAX-RS service. Removed SampleRESTTestBase (no more needed).

Renamed SimpleSOAPWebServiceTest to SimpleWebServiceTest because the test
cover the whole wizard now.

SimpleWSWizard reimplemented to RedDeer and moved to ws.reddeer plugin.
Its superclass was changed to NewWizardDialog which do function of NewSimpleWsWizardAction.
Because of that, the NewSimpleWSWizardAction was removed.