Disable mylyn tests - JBIDE-18000 plus RedDeer refactorings
The integration tests repo fails to compile. This is because mylyn
test defines integration stack Alpha2 repo and that disappeared.
See https://issues.jboss.org/browse/JBIDE-18000 .
Also, mylyn.reddeer fails to compile because of changes introduced
in https://github.com/jboss-reddeer/reddeer/pull/527 . So that is
disabled as well.