WebServices: more RedDeer, sophisticated waiting for jax-rs problems
Implemented sophisticated waiting for JAX-RS Problems that are result of
JAX-RS validation. There are types of waiting that should eliminate false
positive and false negative results.
False positive can occur when Problems
view is empty and we are testing if no problem appeared. In this case an
assertion will wait to ensure that no problem will show as a result of delayed
jax-rs validation.
On the other hand, false negative can occur when we are
testing if there is expected number of problems in Problems view. An assertion
should wait to be sure that Problems view was updated.

Refactored error/warning assertion in RESTfulTestBase.

Added sleep() in order to improve stability until more sophisticated solution
will be implemented (e.g. waiting until the JAX-RS Explorer is updated)

Created JAXRSValidatorPreferencePage in order to reimplment JaxRsValidatorTest
with RedDeer.

Using more RedDeer in RESTfulHelper.