Temporarily disable artifacts from experimental nexus in perf plugin
The build of org.jboss.tools.perf.test.core stopped working
because there is a dependency on some artifacts from
https://repository.jboss.org/nexus/content/repositories/jbosstools-experiments/
and this stopped working - Access Denied.
I reported this at https://issues.jboss.org/browse/ORG-2244 .
The consequence is that part of the perf framework functionality
will not work in tests that consume it (CPU/RAM info etc.).