Add 'Known Problems' section
This fix adds reference to ubuntu related problem with vncserver when it skips
's' chars in input. This leads to test failures, because of test projects
with 's' char in name could not be located in Projects. For example if test
creates projetc with name 'CordovaTestProject' the actula project created in
workspace has name 'CordovaTetProject' and test fails.