JBQA-11696 Unify properties for tests skipping in jbosstools-integration-tests
Until now we used swtbot.test.skip which was set to true in tests/pom.xml.
But it seems this property is redundant - we can simply use skipTests
everywhere and it will be clearer for people outside our team.
skipTests is also the property that is taken by tycho surefire plugin.